+++ app/build.gradle
@@ -1,6 +1,16 @@
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply from: '../git.gradle'
+def versionArray=getArray()
+android {
+    signingConfigs {
+        debug {
+            storeFile file('C:\\MyProject\\CurrencyConversionApp\\keyStore\\android.jks')
+            storePassword '1a2b3c4d'
+            keyAlias 'Converter'
+            keyPassword '1a2b3c4d'
+        }
 
 android {
     compileSdkVersion 30
@@ -19,10 +29,11 @@ android {
         applicationId "com.example.currencyconversionapp"
         minSdkVersion 17
         targetSdkVersion 30
-        versionCode 1
-        versionName "1.0"
+        versionCode generateVersionCode(versionArray)
+        versionName "${versionArray[0]}.${versionArray[1]}.${versionArray[2]}"
 
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        signingConfig signingConfigs.debug
     }
 
     buildTypes {
@@ -41,9 +52,11 @@ android {
     productFlavors{
         free{
             applicationId "com.example.currencyconversionapp.free"
+            dimension 'type'
         }
         paid{
             applicationId "com.example.currencyconversionapp.paid"
+            dimension 'type'
         }
     }
